Fix authentication, still a lot to be done

+++ src/main/java/fr/atlantique/imt/inf211/jobmngt/controller/CandidatController.java
@@ -3,6 +3,7 @@ package fr.atlantique.imt.inf211.jobmngt.controller;
 import fr.atlantique.imt.inf211.jobmngt.entity.AppUser;
 import fr.atlantique.imt.inf211.jobmngt.entity.Candidat;
 import fr.atlantique.imt.inf211.jobmngt.service.CandidatService;
+import fr.atlantique.imt.inf211.jobmngt.service.AuthenticationService;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpSession;
 
@@ -19,6 +20,9 @@ public class CandidatController {
 
     @Autowired
     private CandidatService candidatService;
+
+    @Autowired
+    private AuthenticationService authService;
 
     @GetMapping
     public ModelAndView listCandidates() {
@@ -87,13 +91,10 @@ public class CandidatController {
 
     @GetMapping("/{id}/edit")
     public ModelAndView editCandidateForm(@PathVariable int id, HttpServletRequest request) {
-        HttpSession session = request.getSession();
-        Integer uid = (Integer) session.getAttribute("uid");
-        
-        // Vérification que l'utilisateur est connecté
-        if (uid == null) {
-            return new ModelAndView("redirect:/login");
-        }
+
+        if (!authService.checkCandidatAccess(request.getSession(), id)) {
+            return new ModelAndView("redirect:/error/403");
+        }    
         
         ModelAndView modelAndView = new ModelAndView("candidate/candidateForm");
         Candidat candidate = candidatService.getCandidatById(id);
@@ -109,7 +110,12 @@ public class CandidatController {
     }
 
     @PostMapping("/{id}/edit")
-    public ModelAndView updateCandidate(@PathVariable int id, @ModelAttribute Candidat candidate) {
+    public ModelAndView updateCandidate(@PathVariable int id, @ModelAttribute Candidat candidate, HttpServletRequest request) {
+
+        if (!authService.checkCandidatAccess(request.getSession(), id)) {
+            return new ModelAndView("redirect:/error/403");
+        }
+    
         Candidat existingCandidate = candidatService.getCandidatById(id);
         if (existingCandidate != null) {
             candidate.setIdCandidat(id);
@@ -120,7 +126,10 @@ public class CandidatController {
     }
 
     @GetMapping("/{id}/delete")
-    public ModelAndView deleteCandidate(@PathVariable int id) {
+    public ModelAndView deleteCandidate(@PathVariable int id, HttpServletRequest request) {
+        if (!authService.checkCandidatAccess(request.getSession(), id)) {
+            return new ModelAndView("redirect:/error/403");
+        }    
         candidatService.deleteCandidat(id);
         return new ModelAndView("redirect:/candidates");
     }
